Pull dynamically docker images

+++ docker/build.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
+if ! [ -x "$(command -v envsubst)" ]; then
+  echo "Error: envsubst not installed. To install on Ubuntu run:" >&2
+  echo "sudo apt-get install gettext-base" >&2
+  exit 1
+fi
+
 [[ -f config.sh ]] && source config.sh
+dockerfile=../Dockerfile
 from=../
 to=/tmp/fargate-rails-demo-build/
 rsync -avz --quiet --delete --exclude=.git/ --exclude-from=<(git -C ${from} ls-files --exclude-standard -oi --directory) ${from} ${to}
@@ -10,9 +17,7 @@ mkdir -p $cache_dir
 weekly_id_file=$cache_dir/weekly_id
 [ -f "$weekly_id_file" ] && old_weekly_id=$(<$weekly_id_file)
 if [[ "$old_weekly_id" != "$WEEKLY_ID" ]]; then
-  docker pull docker.io/ruby:${RUBY_VERSION}-slim-buster &
-  docker pull docker.io/node:${NODE_VERSION}-buster-slim &
-  docker pull docker.io/ubuntu:20.04 &
+  grep -oP 'FROM \Kdocker.io/\S+' $dockerfile | envsubst | xargs -I {} docker pull "{}" &
   wait
 fi
 
@@ -21,6 +26,6 @@ DOCKER_BUILDKIT=1 docker build \
   --build-arg RUBY_VERSION=$RUBY_VERSION \
   --build-arg NODE_VERSION=$NODE_VERSION \
   --build-arg BUNDLER_VERSION=$BUNDLER_VERSION \
-  --file ../Dockerfile \
+  --file $dockerfile \
   -t fargate-web-app ${to} &&
 (echo $WEEKLY_ID > $weekly_id_file)
